requirements.txt
flask
python-dotenv
psycopg2-binary
argon2-cffi
python-jose